harvard_algo_trading_project/requirements.txt
# Core trading libraries
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
ccxt>=4.0.0
alpaca-trade-api>=3.0.0

# Technical analysis
ta-lib>=0.4.0
pandas-ta>=0.3.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Machine learning (for analysis, not prediction)
scikit-learn>=1.3.0
scipy>=1.10.0

# Web framework for dashboards
flask>=2.3.0
dash>=2.11.0

# Database
sqlite3
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0

# Logging and monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 